# backend/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"GeminiFrontendConfigResponse": {
				"properties": {
					"configured": {
						"type": "boolean"
					},
					"model": {
						"type": "string"
					},
					"defaults": {
						"properties": {
							"maxOutputTokens": {
								"type": "number",
								"format": "double"
							},
							"temperature": {
								"type": "number",
								"format": "double"
							},
							"maxShortlistSize": {
								"type": "number",
								"format": "double"
							},
							"minShortlistSize": {
								"type": "number",
								"format": "double"
							},
							"shortlistSize": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"maxOutputTokens",
							"temperature",
							"maxShortlistSize",
							"minShortlistSize",
							"shortlistSize"
						],
						"type": "object"
					},
					"contracts": {
						"properties": {
							"candidateFields": {
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							"jobFields": {
								"items": {
									"type": "string"
								},
								"type": "array"
							}
						},
						"required": [
							"candidateFields",
							"jobFields"
						],
						"type": "object"
					},
					"endpoints": {
						"properties": {
							"frontendConfig": {
								"type": "string"
							},
							"screenBatch": {
								"type": "string"
							},
							"screenRun": {
								"type": "string"
							},
							"screenCandidate": {
								"type": "string"
							},
							"generate": {
								"type": "string"
							},
							"health": {
								"type": "string"
							}
						},
						"required": [
							"frontendConfig",
							"screenBatch",
							"screenRun",
							"screenCandidate",
							"generate",
							"health"
						],
						"type": "object"
					}
				},
				"required": [
					"configured",
					"model",
					"defaults",
					"contracts",
					"endpoints"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeminiUsageMetadata": {
				"properties": {
					"promptTokenCount": {
						"type": "number",
						"format": "double"
					},
					"candidatesTokenCount": {
						"type": "number",
						"format": "double"
					},
					"totalTokenCount": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"GeminiGenerateResponse": {
				"properties": {
					"text": {
						"type": "string"
					},
					"model": {
						"type": "string"
					},
					"usage": {
						"$ref": "#/components/schemas/GeminiUsageMetadata"
					}
				},
				"required": [
					"text",
					"model"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeminiGenerateRequest": {
				"properties": {
					"prompt": {
						"type": "string"
					},
					"systemInstruction": {
						"type": "string"
					},
					"temperature": {
						"type": "number",
						"format": "double"
					},
					"maxOutputTokens": {
						"type": "number",
						"format": "double"
					},
					"responseMimeType": {
						"type": "string",
						"enum": [
							"text/plain",
							"application/json"
						]
					},
					"seed": {
						"type": "number",
						"format": "double"
					},
					"topP": {
						"type": "number",
						"format": "double"
					},
					"conversationHistory": {
						"items": {
							"properties": {
								"content": {
									"type": "string"
								},
								"role": {
									"type": "string",
									"enum": [
										"user",
										"assistant"
									]
								}
							},
							"required": [
								"content",
								"role"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"prompt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeminiCriterionAssessment": {
				"properties": {
					"label": {
						"type": "string"
					},
					"weightPct": {
						"type": "number",
						"format": "double"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"weightedScore": {
						"type": "number",
						"format": "double"
					},
					"summary": {
						"type": "string"
					},
					"evidence": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"label",
					"weightPct",
					"score",
					"weightedScore",
					"summary",
					"evidence"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeminiCandidateScreenResponse": {
				"properties": {
					"recommendation": {
						"type": "string",
						"enum": [
							"strong_yes",
							"yes",
							"maybe",
							"no"
						]
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"mustHaveMatchScore": {
						"type": "number",
						"format": "double"
					},
					"dataCompletenessScore": {
						"type": "number",
						"format": "double"
					},
					"summary": {
						"type": "string"
					},
					"strengths": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"concerns": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"evidence": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"criterionAssessments": {
						"items": {
							"$ref": "#/components/schemas/GeminiCriterionAssessment"
						},
						"type": "array"
					},
					"raw": {
						"type": "string"
					},
					"model": {
						"type": "string"
					}
				},
				"required": [
					"recommendation",
					"score",
					"mustHaveMatchScore",
					"dataCompletenessScore",
					"summary",
					"strengths",
					"concerns",
					"evidence",
					"criterionAssessments",
					"raw",
					"model"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeminiWeightCriterion": {
				"properties": {
					"id": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"value": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"label",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeminiRankingCriterion": {
				"properties": {
					"id": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"pct": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"label",
					"pct"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeminiJobContext": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"locationPolicy": {
						"type": "string"
					},
					"employmentType": {
						"type": "string"
					},
					"salaryBand": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"responsibilities": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"mustHaveQualifications": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"niceToHaveQualifications": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"hardSkills": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"coreHardSkills": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"preferredBonusSkills": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"softSkills": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"coreSoftSkills": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"experience": {
						"type": "string"
					},
					"seniorityLevel": {
						"type": "string"
					},
					"educationLevel": {
						"type": "string"
					},
					"weightCriteria": {
						"items": {
							"$ref": "#/components/schemas/GeminiWeightCriterion"
						},
						"type": "array"
					},
					"rankingCriteria": {
						"items": {
							"$ref": "#/components/schemas/GeminiRankingCriterion"
						},
						"type": "array"
					}
				},
				"required": [
					"title"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GeminiCandidateContext": {
				"properties": {
					"id": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"summary": {
						"type": "string"
					},
					"resumeText": {
						"type": "string"
					},
					"importedData": {
						"$ref": "#/components/schemas/GeminiImportedCandidateData"
					}
				},
				"required": [
					"resumeText"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeminiCandidateScreenRequest": {
				"properties": {
					"job": {
						"$ref": "#/components/schemas/GeminiJobContext"
					},
					"candidate": {
						"$ref": "#/components/schemas/GeminiCandidateContext"
					},
					"instructions": {
						"type": "string"
					},
					"temperature": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"job",
					"candidate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeminiFrontendScreeningResult": {
				"properties": {
					"recommendation": {
						"type": "string",
						"enum": [
							"strong_yes",
							"yes",
							"maybe",
							"no"
						]
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"mustHaveMatchScore": {
						"type": "number",
						"format": "double"
					},
					"dataCompletenessScore": {
						"type": "number",
						"format": "double"
					},
					"summary": {
						"type": "string"
					},
					"strengths": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"concerns": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"evidence": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"criterionAssessments": {
						"items": {
							"$ref": "#/components/schemas/GeminiCriterionAssessment"
						},
						"type": "array"
					},
					"raw": {
						"type": "string"
					},
					"model": {
						"type": "string"
					},
					"candidateId": {
						"type": "string"
					},
					"candidateName": {
						"type": "string"
					},
					"shortlisted": {
						"type": "boolean"
					},
					"rank": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"recommendation",
					"score",
					"mustHaveMatchScore",
					"dataCompletenessScore",
					"summary",
					"strengths",
					"concerns",
					"evidence",
					"criterionAssessments",
					"raw",
					"model",
					"shortlisted",
					"rank"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeminiFrontendScreeningRunResponse": {
				"properties": {
					"runName": {
						"type": "string"
					},
					"jobTitle": {
						"type": "string"
					},
					"totalCandidates": {
						"type": "number",
						"format": "double"
					},
					"shortlistSize": {
						"type": "number",
						"format": "double"
					},
					"shortlistedCount": {
						"type": "number",
						"format": "double"
					},
					"model": {
						"type": "string"
					},
					"rankingCriteria": {
						"items": {
							"$ref": "#/components/schemas/GeminiRankingCriterion"
						},
						"type": "array"
					},
					"summary": {
						"type": "string"
					},
					"results": {
						"items": {
							"$ref": "#/components/schemas/GeminiFrontendScreeningResult"
						},
						"type": "array"
					}
				},
				"required": [
					"runName",
					"jobTitle",
					"totalCandidates",
					"shortlistSize",
					"shortlistedCount",
					"model",
					"summary",
					"results"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeminiFrontendScreeningRunRequest": {
				"properties": {
					"runName": {
						"type": "string"
					},
					"job": {
						"$ref": "#/components/schemas/GeminiJobContext"
					},
					"candidates": {
						"items": {
							"$ref": "#/components/schemas/GeminiCandidateContext"
						},
						"type": "array"
					},
					"shortlistSize": {
						"type": "number",
						"format": "double"
					},
					"instructions": {
						"type": "string"
					},
					"temperature": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"runName",
					"job",
					"candidates"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeminiBatchRecommendation": {
				"type": "string",
				"enum": [
					"Strong Reject",
					"Reject",
					"Consider",
					"Shortlist",
					"Strong Shortlist"
				]
			},
			"GeminiBatchScreeningResultEntry": {
				"properties": {
					"candidateRank": {
						"type": "number",
						"format": "double"
					},
					"applicantEmail": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"matchScore": {
						"type": "number",
						"format": "double"
					},
					"confidenceScore": {
						"type": "number",
						"format": "double"
					},
					"skillsScore": {
						"type": "number",
						"format": "double"
					},
					"experienceScore": {
						"type": "number",
						"format": "double"
					},
					"educationScore": {
						"type": "number",
						"format": "double"
					},
					"relevanceScore": {
						"type": "number",
						"format": "double"
					},
					"criterionAssessments": {
						"items": {
							"$ref": "#/components/schemas/GeminiCriterionAssessment"
						},
						"type": "array"
					},
					"criticalRequirementGap": {
						"type": "boolean"
					},
					"strengths": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"gapsOrRisks": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"finalRecommendation": {
						"$ref": "#/components/schemas/GeminiBatchRecommendation"
					},
					"summaryExplanation": {
						"type": "string"
					}
				},
				"required": [
					"candidateRank",
					"applicantEmail",
					"fullName",
					"matchScore",
					"confidenceScore",
					"skillsScore",
					"experienceScore",
					"educationScore",
					"relevanceScore",
					"criticalRequirementGap",
					"strengths",
					"gapsOrRisks",
					"finalRecommendation",
					"summaryExplanation"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GeminiBatchScreeningResponse": {
				"properties": {
					"jobTitle": {
						"type": "string"
					},
					"weightCriteria": {
						"items": {
							"$ref": "#/components/schemas/GeminiWeightCriterion"
						},
						"type": "array"
					},
					"shortlistCount": {
						"type": "number",
						"format": "double"
					},
					"totalApplicants": {
						"type": "number",
						"format": "double"
					},
					"screeningResults": {
						"items": {
							"$ref": "#/components/schemas/GeminiBatchScreeningResultEntry"
						},
						"type": "array"
					},
					"shortlist": {
						"items": {
							"$ref": "#/components/schemas/GeminiBatchShortlistEntry"
						},
						"type": "array"
					},
					"model": {
						"type": "string"
					},
					"meta": {
						"$ref": "#/components/schemas/GeminiBatchScreeningMeta"
					}
				},
				"required": [
					"jobTitle",
					"shortlistCount",
					"totalApplicants",
					"screeningResults",
					"shortlist"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeminiBatchScreeningDbRequest": {
				"properties": {
					"jobId": {
						"type": "string"
					},
					"shortlistCount": {
						"type": "number",
						"format": "double"
					},
					"instructions": {
						"type": "string"
					},
					"temperature": {
						"type": "number",
						"format": "double"
					},
					"applicantIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"applicantEmails": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"filters": {
						"properties": {
							"ingestStatus": {
								"type": "string",
								"enum": [
									"parsed",
									"pending",
									"failed"
								]
							}
						},
						"type": "object"
					}
				},
				"required": [
					"jobId",
					"shortlistCount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ScreeningRunStatus": {
				"type": "string",
				"enum": [
					"queued",
					"running",
					"completed",
					"partial",
					"failed"
				]
			},
			"GeminiBatchScreeningRunStatusResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"jobId": {
						"type": "string"
					},
					"jobTitle": {
						"type": "string"
					},
					"runName": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/ScreeningRunStatus"
					},
					"shortlistCount": {
						"type": "number",
						"format": "double"
					},
					"requestedApplicants": {
						"type": "number",
						"format": "double"
					},
					"processedApplicants": {
						"type": "number",
						"format": "double"
					},
					"scoredApplicants": {
						"type": "number",
						"format": "double"
					},
					"failedApplicants": {
						"type": "number",
						"format": "double"
					},
					"processedChunks": {
						"type": "number",
						"format": "double"
					},
					"failedChunks": {
						"type": "number",
						"format": "double"
					},
					"chunkSize": {
						"type": "number",
						"format": "double"
					},
					"model": {
						"type": "string"
					},
					"failureReasons": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"error": {
						"type": "string"
					},
					"screeningStartedAt": {
						"type": "string"
					},
					"screeningCompletedAt": {
						"type": "string"
					},
					"screeningDurationSeconds": {
						"type": "number",
						"format": "double"
					},
					"savedShortlistId": {
						"type": "string"
					},
					"response": {
						"$ref": "#/components/schemas/GeminiBatchScreeningResponse"
					}
				},
				"required": [
					"id",
					"jobId",
					"jobTitle",
					"runName",
					"status",
					"shortlistCount",
					"requestedApplicants",
					"processedApplicants",
					"scoredApplicants",
					"failedApplicants",
					"processedChunks",
					"failedChunks",
					"chunkSize",
					"failureReasons"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeminiBatchScreeningRunRequest": {
				"properties": {
					"jobId": {
						"type": "string"
					},
					"shortlistCount": {
						"type": "number",
						"format": "double"
					},
					"instructions": {
						"type": "string"
					},
					"temperature": {
						"type": "number",
						"format": "double"
					},
					"applicantIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"applicantEmails": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"filters": {
						"properties": {
							"ingestStatus": {
								"type": "string",
								"enum": [
									"parsed",
									"pending",
									"failed"
								]
							}
						},
						"type": "object"
					},
					"runName": {
						"type": "string"
					}
				},
				"required": [
					"jobId",
					"shortlistCount"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GeminiRecruiterAssistantResponse": {
				"properties": {
					"reply": {
						"type": "string"
					},
					"model": {
						"type": "string"
					},
					"usage": {
						"$ref": "#/components/schemas/GeminiUsageMetadata"
					},
					"contextUsed": {
						"$ref": "#/components/schemas/GeminiRecruiterAssistantContextSummary"
					}
				},
				"required": [
					"reply",
					"model",
					"contextUsed"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeminiRecruiterAssistantRole": {
				"type": "string",
				"enum": [
					"user",
					"assistant"
				]
			},
			"GeminiRecruiterAssistantMessage": {
				"properties": {
					"role": {
						"$ref": "#/components/schemas/GeminiRecruiterAssistantRole"
					},
					"content": {
						"type": "string"
					}
				},
				"required": [
					"role",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeminiRecruiterAssistantRequest": {
				"properties": {
					"message": {
						"type": "string"
					},
					"history": {
						"items": {
							"$ref": "#/components/schemas/GeminiRecruiterAssistantMessage"
						},
						"type": "array"
					},
					"jobId": {
						"type": "string"
					},
					"shortlistId": {
						"type": "string"
					},
					"applicantEmails": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"includeApplicants": {
						"type": "boolean"
					},
					"applicantLimit": {
						"type": "number",
						"format": "double"
					},
					"temperature": {
						"type": "number",
						"format": "double"
					},
					"maxOutputTokens": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"JobImportResponse": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/ImportedJobDraft"
					},
					"warnings": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"data",
					"warnings"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ParseJobLinkRequest": {
				"properties": {
					"url": {
						"type": "string"
					}
				},
				"required": [
					"url"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ParseJobFileRequest": {
				"properties": {
					"fileName": {
						"type": "string"
					},
					"mimeType": {
						"type": "string"
					},
					"base64Data": {
						"type": "string"
					}
				},
				"required": [
					"fileName",
					"base64Data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "backend-api",
		"version": "1.0.0",
		"description": "Simple backend API with Hello World endpoint using TypeScript, tsoa, and Mongoose",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Bikorimana Saveur"
		}
	},
	"paths": {
...
		"/gemini/frontend-config": {
			"get": {
				"operationId": "FrontendConfig",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeminiFrontendConfigResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Gemini"
				],
				"security": [
					{
						"jwt": [
							"recruiter",
							"admin"
						]
					}
				],
				"parameters": []
			}
		},
		"/gemini/generate": {
			"post": {
				"operationId": "Generate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeminiGenerateResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Gemini"
				],
				"security": [
					{
						"jwt": [
							"recruiter",
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GeminiGenerateRequest"
							}
						}
					}
				}
			}
		},
		"/gemini/screen-candidate": {
			"post": {
				"operationId": "ScreenCandidate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeminiCandidateScreenResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Gemini"
				],
				"security": [
					{
						"jwt": [
							"recruiter",
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GeminiCandidateScreenRequest"
							}
						}
					}
				}
			}
		},
		"/gemini/screen-run": {
			"post": {
				"operationId": "ScreenRun",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeminiFrontendScreeningRunResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Gemini"
				],
				"security": [
					{
						"jwt": [
							"recruiter",
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GeminiFrontendScreeningRunRequest"
							}
						}
					}
				}
			}
		},
		"/gemini/screen-batch": {
			"post": {
				"operationId": "ScreenBatch",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeminiBatchScreeningResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Gemini"
				],
				"security": [
					{
						"jwt": [
							"recruiter",
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GeminiBatchScreeningDbRequest"
							}
						}
					}
				}
			}
		},
		"/gemini/screen-batch-runs": {
			"post": {
				"operationId": "StartScreenBatchRun",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"data": {
											"$ref": "#/components/schemas/GeminiBatchScreeningRunStatusResponse"
										}
									},
									"required": [
										"message",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Gemini"
				],
				"security": [
					{
						"jwt": [
							"recruiter",
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GeminiBatchScreeningRunRequest"
							}
						}
					}
				}
			}
		},
		"/gemini/screen-batch-runs/{runId}": {
			"get": {
				"operationId": "GetScreenBatchRun",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"data": {
											"$ref": "#/components/schemas/GeminiBatchScreeningRunStatusResponse"
										}
									},
									"required": [
										"message",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Gemini"
				],
				"security": [
					{
						"jwt": [
							"recruiter",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "runId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/gemini/assistant": {
			"post": {
				"operationId": "Assistant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeminiRecruiterAssistantResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Gemini"
				],
				"security": [
					{
						"jwt": [
							"recruiter",
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GeminiRecruiterAssistantRequest"
							}
						}
					}
				}
			}
		},
		"/gemini/parse-job-link": {
			"post": {
				"operationId": "ParseJobLink",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JobImportResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Gemini"
				],
				"security": [
					{
						"jwt": [
							"recruiter",
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ParseJobLinkRequest"
							}
						}
					}
				}
			}
		},
		"/gemini/parse-job-file": {
			"post": {
				"operationId": "ParseJobFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JobImportResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Gemini"
				],
				"security": [
					{
						"jwt": [
							"recruiter",
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ParseJobFileRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}